# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MiAu API REST",
    "description": "\n## MiAu — API REST de Gestão Petshop\n\nBackend **FastAPI** integrado ao banco **MariaDB/MySQL** (`miau_db`).\n\n### Autenticação JWT\n\n1. `POST /auth/login` — envie `username` e `password` (form-urlencoded)\n2. Copie o `access_token` retornado\n3. Clique em **Authorize** e informe: `Bearer {token}`\n4. Rotas `/api/*` exigem token válido\n\n**Usuário padrão:** `ShardCadu` / `cadu123`\n\n### Tabelas do banco\n\n| Tabela | Endpoints |\n|--------|-----------|\n| `usuarios` | `/auth/login`, `/auth/register`, `/auth/me` |\n| `tutores` | `/api/tutores` |\n| `pets` | `/api/pets` |\n| `servicos` | `/api/servicos` |\n| `agendamentos` | `/api/agendamentos` |\n| `avisos` | `/api/avisos` |\n\nDocumentação completa: repositório → `docs/API-Swagger.md`\n",
    "version": "1.0.0",
    "contact": {
      "name": "Carlos Eduardo (Cadu)",
      "url": "https://github.com/Dev-Carlos-Alves/MiAu-"
    },
    "license": {
      "name": "Projeto Acadêmico — Henning 2026.1"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local (python app.py)"
    },
    {
      "url": "https://mi-au.vercel.app",
      "description": "Producao Vercel"
    }
  ],
  "paths": {
...
    "/api/tutores": {
      "get": {
        "tags": [
          "Tutores"
        ],
        "summary": "Listar tutores",
        "description": "Retorna todos os registros da tabela `tutores`.",
        "operationId": "read_tutores_api_tutores_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Tutor"
                  },
                  "type": "array",
                  "title": "Response Read Tutores Api Tutores Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Tutores"
        ],
        "summary": "Cadastrar tutor",
        "description": "Insere registro na tabela `tutores`.",
        "operationId": "create_tutor_api_tutores_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tutor"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/tutores/{tutor_id}": {
      "put": {
        "tags": [
          "Tutores"
        ],
        "summary": "Atualizar tutor",
        "description": "Atualiza tutor existente por ID.",
        "operationId": "update_tutor_api_tutores__tutor_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tutor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tutor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tutores"
        ],
        "summary": "Excluir tutor",
        "description": "Remove tutor por ID (pets vinculados são excluídos em cascata).",
        "operationId": "delete_tutor_api_tutores__tutor_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tutor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tutor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pets": {
      "get": {
        "tags": [
          "Pets"
        ],
        "summary": "Listar pets",
        "description": "Retorna todos os registros da tabela `pets`.",
        "operationId": "read_pets_api_pets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Pet"
                  },
                  "type": "array",
                  "title": "Response Read Pets Api Pets Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Pets"
        ],
        "summary": "Cadastrar pet",
        "description": "Insere registro na tabela `pets`. Espécie: Cachorro ou Gato.",
        "operationId": "create_pet_api_pets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/pets/{pet_id}": {
      "put": {
        "tags": [
          "Pets"
        ],
        "summary": "Atualizar pet",
        "description": "Atualiza pet existente por ID.",
        "operationId": "update_pet_api_pets__pet_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PetCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pets"
        ],
        "summary": "Excluir pet",
        "description": "Remove pet por ID.",
        "operationId": "delete_pet_api_pets__pet_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/servicos": {
      "get": {
        "tags": [
          "Serviços"
        ],
        "summary": "Listar serviços",
        "description": "Retorna catálogo da tabela `servicos`.",
        "operationId": "read_servicos_api_servicos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Servico"
                  },
                  "type": "array",
                  "title": "Response Read Servicos Api Servicos Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Serviços"
        ],
        "summary": "Cadastrar serviço",
        "description": "Insere novo serviço com nome, descrição e preço.",
        "operationId": "create_servico_api_servicos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServicoCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Servico"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/servicos/{id}": {
      "put": {
        "tags": [
          "Serviços"
        ],
        "summary": "Atualizar serviço",
        "description": "Atualiza serviço existente por ID.",
        "operationId": "update_servico_api_servicos__id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServicoCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Servico"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Serviços"
        ],
        "summary": "Excluir serviço",
        "description": "Remove serviço por ID.",
        "operationId": "delete_servico_api_servicos__id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agendamentos": {
      "get": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "Listar agendamentos",
        "description": "Retorna todos os agendamentos com FKs de tutor, pet e serviço.",
        "operationId": "read_agendamentos_api_agendamentos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Agendamento"
                  },
                  "type": "array",
                  "title": "Response Read Agendamentos Api Agendamentos Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "Criar agendamento",
        "description": "Insere agendamento cruzando tutor_id, pet_id e servico_id.",
        "operationId": "create_agendamento_api_agendamentos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgendamentoCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agendamento"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/agendamentos/{id}": {
      "put": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "Atualizar agendamento",
        "description": "Atualiza agendamento existente por ID.",
        "operationId": "update_agendamento_api_agendamentos__id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgendamentoCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agendamento"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Agendamentos"
        ],
        "summary": "Excluir agendamento",
        "description": "Remove agendamento por ID.",
        "operationId": "delete_agendamento_api_agendamentos__id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/avisos": {
      "get": {
        "tags": [
          "Avisos"
        ],
        "summary": "Listar avisos do mural",
        "description": "Retorna avisos ordenados por `data_criacao` (mais recentes primeiro).",
        "operationId": "read_avisos_api_avisos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Aviso"
                  },
                  "type": "array",
                  "title": "Response Read Avisos Api Avisos Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Avisos"
        ],
        "summary": "Publicar aviso",
        "description": "Insere aviso na tabela `avisos`. Tipo: Urgente, Aviso ou Lembrete.",
        "operationId": "create_aviso_api_avisos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvisoCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Aviso"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/avisos/{aviso_id}": {
      "delete": {
        "tags": [
          "Avisos"
        ],
        "summary": "Excluir aviso",
        "description": "Remove aviso do mural por ID.",
        "operationId": "delete_aviso_api_avisos__aviso_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "aviso_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Aviso Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Tutor": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 100,
            "title": "Nome",
            "description": "Nome completo do tutor",
            "example": "João Silva"
          },
          "contato": {
            "type": "string",
            "maxLength": 50,
            "title": "Contato",
            "description": "E-mail ou contato principal",
            "example": "joao@email.com"
          },
          "endereco": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Endereco",
            "description": "Endereço residencial",
            "example": "Rua das Flores, 123"
          },
          "telefone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefone",
            "description": "Telefone com DDD",
            "example": "(11) 98888-1111"
          },
          "foto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foto",
            "description": "Foto do tutor em base64 (data URL)"
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "ID gerado pelo banco (PK)",
            "example": 1
          }
        },
        "type": "object",
        "required": [
          "nome",
          "contato",
          "id"
        ],
        "title": "Tutor"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "description": "Autenticação JWT. No Authorize, use username `ShardCadu` e password `cadu123`. Deixe client_id e client_secret vazios. Rotas /api/* exigem token."
      },
      "BearerJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Alternativa manual: faça POST /auth/login, copie access_token e cole aqui (apenas o token, sem a palavra Bearer)."
      }
    }
  },
...
}